fix: snapshot service refactoring: first pass
