+++ packages/services/pkg/snapshot/constants.go
@@ -2,9 +2,3 @@ package snapshot
 
 // SnapshotDir is the directory name for where the ECS snapshots are stored.
 const SnapshotDir string = "snapshots"
-
-// SerializedStateFilename is the name for the snapshot binary of ECS State.
-const SerializedStateFilename string = "./snapshots/SerializedECSState"
-
-// SerializedWorldsFilename is the name for the snapshot binary of Worlds.
-const SerializedWorldsFilename string = "./snapshots/SerializedWorlds"
